requirements.txt
tensorflow==2.13.0
keras==2.13.1
streamlit
numpy
pillow
gdown